requirements.txt
Flask
Flask-Cors
PyJWT
mysql-connector-python
